only run geo-type doc-tests when when geo-types enabled, but always show them.
And fixed a broken ignored doctest in lib.rs

Admittedly, this is ugly, but:

Using geo-types is a good demonstration of our ToWkt trait, but if
geo-types is disabled, the example will fail.

We could ```ignore the doc blocks, but then the examples are prone to
going stale (as happened to the broken test in lib.rs).