FromWkt trait for reading WKT without exposing the user to the intermediate representation.
Note that this isn't a performance change. It's about (hopefully) making
the library easier to use.

This is a corollary to https://github.com/georust/wkt/pull/89, but for
reading WKT, rather than writing. As we discussed there, probably there
is no reason for the user to care about the Wkt struct.

Note that the intermediate representation is still used (for now!), but
the user is no longer required to interact with it.

The road is open though for having a direct translation from Wkt text to
the geo-types (or whatever) represenation (see geozero for inspiration).